unification id type into int

+++ src/main/java/pl/kosieradzki/siitask/repo/EventRepo.java
@@ -3,5 +3,5 @@ package pl.kosieradzki.siitask.repo;
 import org.springframework.data.jpa.repository.JpaRepository;
 import pl.kosieradzki.siitask.model.Event;
 
-public interface EventRepo extends JpaRepository<Event, Long> {
+public interface EventRepo extends JpaRepository<Event, Integer> {
 }
